+++ weather-app/Network/Weather/ResponseModel/CoordinatesResponseModel.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 struct CoordinatesResponseModel: Codable {
-    let lon: Float
-    let lat: Float
+    let lon: Double
+    let lat: Double
 }
